Fixed #125: adding title to ZIM entries
Because ted2zim's UI is somewhat multilang, video pages contains a list of the
retrieved titles for all the wanted languages.
A piece of JS uses the URL query language to set the <title>'s accordingly but the
HTML source had no title, resulting in empty Title on ZIM entries (ted2zim is still
using zimwriterfs-like mode).
Without changing this behavior, this sets the <title> to the *main* language's one.
This language is chosen on --locale or defaults to video's default or English.
Suggestions do work now.